requirements.txt
fastapi
uvicorn
google-genai
sqlalchemy
pydantic
python-dotenv
